chore(cli): move setup back to main command
