+++ actions/integration-test/steps/setup.sh
@@ -2,7 +2,7 @@
 # IP address in environment "INPUT_IPADDRESS"
 
 if [[ $INPUT_CLI != '' ]]; then
-	${INPUT_CLI_BIN_PATH} project setup "${INPUT_IPADDRESS}" -c "/tmp/.stackhead-cli.yml" -v
+	${INPUT_CLI_BIN_PATH} setup "${INPUT_IPADDRESS}" -c "/tmp/.stackhead-cli.yml" -v
 else
 	INVENTORY_PATH="${GITHUB_ACTION_PATH}/inventory.yml"
 	sed -e "s/\${ipaddress}/${INPUT_IPADDRESS}/" -e "s/\${webserver}/${INPUT_WEBSERVER}/" -e "s/\${container}/${INPUT_CONTAINER}/" -e "s/\${plugins}/${INPUT_PLUGINS}/" "${GITHUB_ACTION_PATH}/inventory.dist.yml" >"${INVENTORY_PATH}"
